refactor: 💡 move exceptions into separate packages
