+++ apps/java-api/src/main/java/org/scoalaonline/api/exception/lectureMaterial/LectureMaterialInvalidDocumentException.java
@@ -1,9 +1,10 @@
-package org.scoalaonline.api.exception;
+package org.scoalaonline.api.exception.lectureMaterial;
 
 public class LectureMaterialInvalidDocumentException extends Exception{
   public LectureMaterialInvalidDocumentException(){
   }
+
   public LectureMaterialInvalidDocumentException(String message){
     super(message);
   }
-  }
+}
