+++ apps/java-api/src/main/java/org/scoalaonline/api/exception/lectureMaterial/LectureMaterialNotFoundException.java
@@ -1,8 +1,9 @@
-package org.scoalaonline.api.exception;
+package org.scoalaonline.api.exception.lectureMaterial;
 
 public class LectureMaterialNotFoundException extends Exception{
   public LectureMaterialNotFoundException(){
   }
+
   public LectureMaterialNotFoundException(String message){
     super(message);
   }
