+++ apps/java-api/src/main/java/org/scoalaonline/api/exception/subject/SubjectInvalidValueException.java
@@ -1,8 +1,9 @@
-package org.scoalaonline.api.exception;
+package org.scoalaonline.api.exception.subject;
 
 /**
  * Exception class for when the Subject's value is invalid.
  */
+
 public class SubjectInvalidValueException extends Exception {
   public SubjectInvalidValueException() {
   }
