+++ apps/java-api/src/main/java/org/scoalaonline/api/exception/subject/SubjectNotFoundException.java
@@ -1,8 +1,9 @@
-package org.scoalaonline.api.exception;
+package org.scoalaonline.api.exception.subject;
 
 /**
  * Exception class for when the Subject entity cannot be found.
  */
+
 public class SubjectNotFoundException extends Exception {
 
   public SubjectNotFoundException() {
